simple_snowplow/requirements.txt
aiochclient[speedups]==2.1.0
aiofiles==0.8.0
aiohttp==3.8.1
brotli==1.0.9
brotli-asgi==1.1.0
dynaconf==3.1.9
elastic-apm==6.10.1
inflection==0.5.1
orjson==3.7.8
PyYAML==6.0
requests==2.28.1
starlette_exporter==0.13.0
ua-parser==0.15.0
user-agents==2.2.0
